Min2

+++ controller/opendict_words.js
@@ -319,16 +319,16 @@ function deleteWord() {
       }
 
       // 이미 나만의 단어장에 단어를 저장한 사용자가 있는 경우 오픈사전 단어장에서 남아 있는 단어 뜻이 하나 밖에 없는 상황에서는 삭제 불가
-      // const findMydictWord = await Mydict.find({ scriptId, word });
-      // const findOpendictWord = await Opendict.find({ scriptId, word });
+      const findMydictWord = await Mydict.find({ scriptId, word });
+      const findOpendictWord = await Opendict.find({ scriptId, word });
 
-      // if (findMydictWord && findOpendictWord.length === 1) {
-      //   return res.json({
-      //     ok: false,
-      //     errorMessage:
-      //       "이미 나만의 단어장에 단어를 저장한 사용자가 있어 삭제할 수 없습니다.",
-      //   });
-      // }
+      if (findMydictWord && findOpendictWord.length === 1) {
+        return res.json({
+          ok: false,
+          errorMessage:
+            "이미 나만의 단어장에 단어를 저장한 사용자가 있어 삭제할 수 없습니다.",
+        });
+      }
 
       await Opendict.deleteOne({ scriptId, wordId });
